Fix failures for Github Actions
