+++ contrib/storage-cassandra/pom.xml
@@ -93,4 +93,15 @@
     </dependency>
   </dependencies>
 
+  <build>
+    <plugins>
+      <plugin>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <configuration>
+          <forkCount combine.self="override">1</forkCount>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+
 </project>
