+++ contrib/storage-cassandra/src/test/java/org/apache/drill/exec/store/cassandra/CassandraPlanTest.java
@@ -17,62 +17,9 @@
  */
 package org.apache.drill.exec.store.cassandra;
 
-import com.datastax.driver.core.Cluster;
-import com.datastax.driver.core.Session;
-import com.github.nosan.embedded.cassandra.Cassandra;
-import com.github.nosan.embedded.cassandra.CassandraBuilder;
-import com.github.nosan.embedded.cassandra.Settings;
-import com.github.nosan.embedded.cassandra.cql.CqlScript;
-import org.apache.drill.test.ClusterFixture;
-import org.apache.drill.test.ClusterTest;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class CassandraPlanTest extends ClusterTest {
-
-  private static Cassandra cassandra;
-
-  @BeforeClass
-  public static void init() throws Exception {
-    startCluster(ClusterFixture.builder(dirTestWatcher));
-
-    startCassandra();
-
-    CassandraStorageConfig config = new CassandraStorageConfig(
-        cassandra.getSettings().getAddress().getHostAddress(),
-        cassandra.getSettings().getPort(),
-        null,
-        null);
-    config.setEnabled(true);
-    cluster.defineStoragePlugin("cassandra", config);
-
-    prepareData();
-  }
-
-  @AfterClass
-  public static void cleanUp() {
-    cassandra.stop();
-  }
-
-  private static void startCassandra() {
-    cassandra = new CassandraBuilder().build();
-    cassandra.start();
-  }
-
-  private static void prepareData() {
-    Settings settings = cassandra.getSettings();
-
-    try (Cluster cluster = Cluster.builder()
-        .addContactPoints(settings.getAddress())
-        .withPort(settings.getPort())
-        .withoutMetrics()
-        .withoutJMXReporting()
-        .build()) {
-      Session session = cluster.connect();
-      CqlScript.ofClassPath("queries.cql").forEachStatement(session::execute);
-    }
-  }
+public class CassandraPlanTest extends BaseCassandraTest {
 
   @Test
   public void testProjectPushDown() throws Exception {
